# perf: Fetching exchange rate on every line item slows down PO